refactor(marketing): split features-section into per-card files
The 1,261-line features-section.tsx held a shared scroll hook plus six
self-contained card components. Extract each into its own file under
components/features/ so the section file is just the orchestrator (89 lines).

No behavior change — code moved verbatim:
- features/use-scroll-motion.ts — useScrollMotion hook + Motion type
- features/database-connection-card.tsx
- features/region-globe-card.tsx
- features/native-performance-card.tsx
- features/query-history-card.tsx
- features/docker-containers-card.tsx
- features/schema-diagram-card.tsx

Verified with typecheck and a production build.
